Add support for broadcasting to handlers based on routing data
Summary:
This is the next step in RTMP proxy. Once a connection gets hashed onto
a particular worker thread based on the routing data, there needs to be
logic to make sure there is a unique upstream connection for any given
routing data that broadcasts bytes flowing from the upstream to all the
pipelines that have subscribed to it.

This fits the rx model. I have made use of Observer, but instead of
using Observable, I've implemented my own BroadcastHandler because
1) Observable uses plenty of locking primitives which I don't want
since all this is thread-local
2) wangle/rx/README warns against using it in prod

I've also implemented a BroadcastPool (which is supposed to be used
as a folly::ThreadLocalPtr) that maintains a map of routingData ->
broadcast pipelines. This also holds the logic to create new upstream
connections (using ClientBootstrap) and broadcast pipelines as
necessary.

The way this would fit into the bigger picture is, the child pipeline
created by AcceptRoutingHandler will contain a new ObservingHandler
(diff coming) that implements Observer and holds a thread-local
BroadcastPool. It'll invoke BroadcastPool::getHandler to get a
BroadcastHandler and susbscribe to it and stream the data downstream.

All this is very generic and there's no RTMP logic here. I'd like to
move this into wangle if you guys are good with that. Longer term, it'd
be cool to integrate more of rx into this.

Test Plan: Unit tests

Reviewed By: davejwatson@fb.com

Subscribers: jsedgwick, bmatheny, praveenr, afrind, fugalh

FB internal diff: D2190346

Tasks: 6804728

Signature: t1:2190346:1435768678:29a136adc4610b5606670b49ab64b4ed8aa1b631