Templatize ObservingHandler and add MessageToByteEncoder
Summary:
Previously ObservingPipeline was a DefaultPipeline that always took
IOBuf in the write path. This diff refactors it to take an
arbitrary message type that will be braodcasted by the BraodcastPipeline.
ObservingHandler will listen to this message (via onNext() callback)
and pass it over to an outbound MessageToByteEncoder which encodes this
into an IOBuf.

The MessageToByteEncoder is an inverse of ByteToMessageDecoder and
follows Netty's terminology.

Test Plan: Unit tests, hacked up accept_server and ran it.

Reviewed By: noamler@fb.com

Subscribers: fugalh, jsedgwick, yfeldblum

FB internal diff: D2298218

Tasks: 6804728

Signature: t1:2298218:1438354380:4a80506ddc2629395ef8073878f20e1eaee97dbd