Fix race in IOThreadPoolExecutor
Summary: startupBaton is signaled before the loop is running. This diff signals the
baton as the first thing the event base does _after_ starting.

When creating an executor as a local variable in a function, and using its
event bases to make thrift calls, the event bases may not be running by
the time the calls are made.

Reviewed By: yfeldblum

Differential Revision: D2385012

fb-gh-sync-id: f691935efd019260531bc57f7e8d928274267287