Make getSSLSession const
Summary: Make the get method const for ssl session.
This is needed if we need to do more efficient locking for
getSSLSession.

To do this I had to remove removeSSLSession from SSLPersistentCache.
I'm going to rely on the evicting cache implementation to evict
stale entries from the cache when needed / a new session will replace
the stale session almost immediately when there is a cache miss due
to a full handshake setting a ticket.

Reviewed By: mzlee

Differential Revision: D2388301

fb-gh-sync-id: 712f7bf05b65e7067411e090edc40ae46b92ddeb