move Pipeline ownership model from DelayedDestruction to shared_ptr
Summary: I hate almost everything about this diff, but if we want to support people firing pipeline events from multiple threads (which has been handy especially with EventBaseHandler in tow), then we need something like it. DelayedDestruction is not threadsafe and one client sees crashes semiregularly in prod - see the task.
The alternative is just to mandate that folks switch to the IO thread themselves before manipulating the pipeline. I am not fully opposed to such a mandate if we feel it's cleaner than the solution here (it would definitely be more performant)

Reviewed By: djwatson

Differential Revision: D2365748

fb-gh-sync-id: cf53ec1df72f6cfcdf403366da999a6b6fdbdba0