improve wangle::Codel time units
Summary: Pretty things up a bit wrt time units in wangle::Codel.
Fix a bug in `getLoad` - it was off by a factor of ten because it was accidentally dividing microseconds by milliseconds and then not multiplying by 100 to get a percentage.

WHAT WILL THIS BREAK? If fixing `getLoad` might break your code speak up now.

Reviewed By: yfeldblum

Differential Revision: D2406731

fb-gh-sync-id: 7d9224062e686194422b742311e6e05645ec7f9c