Fix thrift server shutdown hanging race condition in ServerBootstrap
Summary: [wangle] Fix thrift server shutdown hanging race condition in ServerBootstrap

ServerBootstrap uses folly::SharedMutexWritePriority (aliased to folly::SharedMutex). This introduces a deadlock when the lock is acquired in the order R->W->R, as the writer will block the second reader from acquiring the lock, which the first reader is waiting for.

Specifically,
- ServerWorkerPool::threadStopped takes a reader lock, and invokes worker->second->getEventBase()->runImmediatelyOrRunInEventBaseThreadAndWait, which requires the EventBase thread to execute the requested work;
- A previously shut down I/O thread now tries to acquire the lock in writer mode (the worked requested in worker->second->getEventBase()->runAfterDrain)
- The EventBase thread being waited by threadStopped runs ServerWorkerPool::forEachWorker and tries to acquire the same lock in reader mode, resulting in a dead lock.

The fix is to switch to folly::SharedMutexReadPriority, which disallows any writer to enter before all readers can finish, thus preventing the above deadlock situation.

Since the writer only enters the picture during shutdown, writer starvation should not be a problem.

Reviewed By: yfeldblum

Differential Revision: D2520826

fb-gh-sync-id: 0d77702b941d23dd31f0231434935c542da0b1fa