Fix use-after-free if ObservingHandler is deleted before upstream connect
Summary: ObservingHandler initiates an upstream connect request by invoking
broadcastPool_->getHandler(). If the ObservingHandler is deleted as a result of
client errors, then when the connect request returns, we'll be accessing freed
memory.

As a side effect of this fix, we could end up leaking upstream connections if
the BroadcastHandler is not cleaned up properly. This diff addresses that as
well.

Reviewed By: federico

Differential Revision: D2610547

fb-gh-sync-id: 1c3816360437c2878729aafeed8e925d21a0af8e