Fix flaky BootstrapTest
Summary:
UDPClientServerTest started being flaky due to a race-condition and was failing
with a CHECK failure CHECK(connections == 1). This is because when the CHECK is
executed, the acceptor thread may not have processed the client connection (and
connections wouldn't have been incremented). The acceptor thread needs to be
joined after shutting down the server before checking.

Also, converted CHECKs to EXPECT_EQ.

Reviewed By: yfeldblum

Differential Revision: D2905076

fb-gh-sync-id: e9c50a55c9b20f35a507d26c5536ea353a9a7af4