Pass exceptions while parsing routing data to AcceptPipeline
Summary:
When an exception occurs while parsing RoutingData from the socket from within
the RoutingPipeline, we invoke the onError() callback on AcceptRoutingHandler
to delete the RoutingPipline. This diff updates the callback to take the
exception_wrapper as well and pass it to all handlers of the AcceptPipeline.
AcceptPipeline can have stats handlers that export counters based on these
exceptions.

Reviewed By: yfeldblum

Differential Revision: D2905301

fb-gh-sync-id: b2422f58e1535ff48d35945e2b8ebe694cc2ef66