ConnectionManager: handle new connections during CLOSE_WHEN_IDLE state
Summary:The previous logic assumed that closeWhenIdle would not destroy the connection, or that all callers to addConnection did not access the connection afterwards.  This moves closeWhenIdle into a loop callback, so addConnection will never invoke it directly.  I also handle the case where there was no-idle grace, so the CM is skipping notifyPendingShutdown.

I guess an alternate approach would be to chase down call sites and make them hold a DestructorGuard while calling addConnection.

Reviewed By: viswanathgs

Differential Revision: D3049913

fb-gh-sync-id: aec1986f06f6b80c5b541e995909af3109c0f5ea
shipit-source-id: aec1986f06f6b80c5b541e995909af3109c0f5ea