Changes for OS X support
Summary:
I have made a few changes in order to get Wangle to build on OS X. There were 4 different problems which I've fixed in 4 different commits. The commit messages should be self-explanatory. The c++14 requirement might not be the best option to solve the compilation issue depending on the project's objective. Maybe the methods that relied on return type deduction should just specify their return type explicitly so as to remain c++11 compatible. I leave that to discussion.
Closes https://github.com/facebook/wangle/pull/47

Differential Revision: D3239979

Pulled By: elliottneilclark

fbshipit-source-id: 9759c9621bb8f1d8e798ff67eb838dce9112b7a2