Improve correctness of (Priority)LifoSemMPMCQueue
Summary:
`(Priority)LifoSemMPMCQueue` relies on the invariant that the
semaphore's value is greater or equal than the number of enqueued
elements, so that it can make progress.

The current implementation has a subtle bug that can break this
invariant, stalling the queue indefinitely.

T0:
Call `take()`, `sem_` enters waiting state.

T1:
Call `add()`: acquire push ticket 0, do not complete the write.

T2:
Call `add()`: acquire push ticket 1, complete the write, post `sem_`

T0:
`sem_` wakes up the thread, `read()` tries pop ticket 0, but
the element cannot be dequeued yet, so returns `false`.

Then we have decreased the semaphore value without dequeuing anything.

This is extremely unlikely to happen in practice, because of the
optimistic dequeuing strategy: as soon as another element is enqueued,
multiple elements can be dequeued with a single semaphore handoff.

Still, better fix this for correctness: `readIfNotEmpty()` is
guaranteed to return `true` if there's a pop ticket available (and it
will wait until the write is completed).

Reviewed By: philippv

Differential Revision: D3527722

fbshipit-source-id: 1dcd9791552448703b743803134ceb531cbcd622