Add SSL rejection
Summary:
If SSL traffic comes in on a plaintext port,
it is desirable to reject it. In case we don't
reject it immediately, the client will only
know when it times out which is not great.

This adds a manager which peeks at a plaintext
thrift request and if it is ssl, sends back
an alert rejecting it.

Openssl does not support a method which allows
us to send an alert. Since this is just a
plaintext alert, we will simply construct it
and send it back to the client before dropping
the connection.

Reviewed By: siyengar

Differential Revision: D3528021

fbshipit-source-id: cf3e5efc319b4327138905c5e46f7acf681b8697