Make Cpp2Connection keep a shared_ptr to Cpp2Worker
Summary:
This removes the need to keep Cpp2Worker in the workers set and rely on EventBase::runAfterDrain to evict it. This doesn't work well if same IO thread pool is being re-used with many different worker pools.

As a follow up this allows killing EventBase::runAfterDrain, since this is the only place where it is used.

Reviewed By: yfeldblum, djwatson

Differential Revision: D3824227

fbshipit-source-id: 0da0591fc31c982d8b67d9903aaa640ddbc2e7e2