enforce lock on ThreadPoolExecutor
Summary:
The spin lock `threadListLock_` which suppose to protect the thread list manipulation was not actually held during critical section.
Only a temporary holder object was constructed on lock, it is destructed right away and the critical section after was not actually protected.
If we call stop/join/setNum to reduce the number of threads in the executor while sill serving traffic, it causes a race.

I run into this problem that an "about to delete" thread is chosen for a new task to be executed on, the `add` function accessed the eventbase pointer in `IOExecutor` which is destructed and the pointer set to nullptr and causing a seg fault.

This diff is intended to enforce those locks, however, there is a little problem with it, it wasn't a problem before because the code were actually lockless.
I added comments on the newly added function, basically, hold the lock, remove the thread from available thread list, release the lock, then join the threads.

Reviewed By: djwatson

Differential Revision: D4307911

fbshipit-source-id: 5c86f8350cf6b67d637a05ac35114515b4f4e323