Wangle: Fix possible deadlock on shutdown of ThreadPoolExecutor
Summary:
DylanZA found that it is still possible to go deadlock with D4307911, after`shouldRun` is set to false

  In wangle/concurrent/IOThreadPoolExecutor.cpp
  ioThread->shouldRun = false;

and `loopForever` returned, it should not go inside the loop again.

  while (ioThread->shouldRun) {
    ioThread->eventBase->loopForever();
  }

However, loopForever may not be able to return because some future is trying to call `add` on `IOThreadPoolExecutor` on destruction.

CPUThreadPoolExecutor is tricky, cause the old way of doing it is that stopThreads will put sential tasks on the taskQueue, whichever thread receives them will be terminated, it is not very trivial to pre-decide which one thread to terminate, this diff will keep it this way, but delegate `remove thread from threadList_` to derived class as well.

This part of code in `wangle/concurrent/ThreadPoolExecutor.cpp`

  for (size_t i = 0; i < n; i++) {
    auto thread = stoppedThreads_.take();
    threadList_.remove(thread);
    waitingForJoinThreads_.add(thread);
  }

is used to remove all the stoppedThread from threadList_, we delegate this part of work to derived class so it can be asynchronized like `CPUThreadPoolExecutor`, and synchronized like `IOThreadPoolExecutor`, changes in the derived classes are just executing this removal.

Reviewed By: djwatson, DylanZA

Differential Revision: D4315295

fbshipit-source-id: 0f8af0d4141da93a2ce90ed7df49a80433c41f49