Fix ownership of TLSTicketKeyManager and SSLSessionCacheManager
Summary:
TLSTicketKeyManager can get destroyed from underneath the SSLContext when certificates are rotated.  This can happen when a rotation occurs while a new connection is being established.  The AsyncSSLSocket keeps a shared_ptr to the context (keeping it alive) but the SSLContexts (owned by SSLContextManager) get cleared and the openssl callbacks setup on the SSL Context point to deleted memory.

The same pattern is used w/ SSLSessionCacheManager as well, so fix that here too.

Reviewed By: knekritz

Differential Revision: D4377694

fbshipit-source-id: a6e6c75585feff2ff37f7c283c4c07d01eaad63e