Fix static linking
Summary:
I was trying to compile a static-only version of this library that links
with static versions of the dependencies. A few things seem to be
missing from the CMake configuration to make this possible.

The extra find_path commands allow for the (admittedly uncommon)
situation where every dependency is installed in a separate location.
Closes https://github.com/facebook/wangle/pull/90

Reviewed By: eduardo-elizondo

Differential Revision: D4720372

Pulled By: anirudhvr

fbshipit-source-id: 895b797779249aa3581aa854f02739bbf000ccab