Remove acceptor dependency in handshake helpers
Summary:
This diff alters the AcceptorHandshakeManager::getHelper interface by no longer
passing a reference to a wangle::Acceptor*.

Previously, handshake helper implementations would use the acceptor reference
to update SSL statistics. Now, updateSSLStats is called automatically by
AcceptorHandshakeManager on connectionReady() and connectionError() if
an SSL error code is supplied.

The motiviation for removing references to acceptors inside of helpers is
primarily due to upcoming changes that will allow handshakes to be performed
under a separate eventbase thread. Having a reference to an acceptor makes it
easy to cause data races (wangle::Acceptor was not designed under the assumption
that it would be used from multiple threads).

Reviewed By: knekritz

Differential Revision: D5628405

fbshipit-source-id: 8d80f34ab04ca50cd87df36f8832207940dca4a4