Detach handlers from pipelines to allow reuse
Summary: The attachCount_ counter guards against getting the context if you have attached the handler to multiple pipelines. However, it is never decreased and so it effectively prevents reuse of a handler - adding detachContext when the pipeline is detached returns the handler to it's original state, which can be used again.

Reviewed By: philipjameson

Differential Revision: D5664116

fbshipit-source-id: 89499676bcfa097637c9aa13df63cb92e5b6fa70