remove rx dependency in ThreadPoolExecutor
Summary:
Without the rx dep wangle/concurrent/ can be moved to folly
Rx functionality is just replaced by a simple, permanent callbacks list
Most of the complexity in rx comes from properly handling reentrancy, e.g. subscribing, unsubscribing, and publishing
in the course of a callback. Now you can't subscribe in a callback (runtime_error thrown), which probably won't ever happen in this case.
As for unsubscribing, it's done away with entirely. If the executor outlives any data on which your callbacks depend, well... poo-poo to you, figure it out yourself
This is plenty good enough for now as this feature is mostly unused

Reviewed By: yfeldblum

Differential Revision: D5686103

fbshipit-source-id: 72cfdfef5b6f6ae4f0617a3dfb6a0725462dc088