reload certs on CA change
Summary:
Thrift servers should reload SSL configuration if cert, key, or CA path change.  TLSCredProcessor had some deficiencies in that it tracked only a single file to trigger the callback.  This augments that class.

Additionally we remove the watchCertForChanges method on ThriftServer and set up watching when a config is set.

Reviewed By: mingtaoy

Differential Revision: D6501821

fbshipit-source-id: fb4f723af644361b3a87354ca124cace94ad04dc