Allow access to tcpinfo on macOS, using tcp_connection_info
Summary:
On macOS, the `tcp_connection_info` struct provides almost the same information as the `tcp_info` struct on Linux and FreeBSD. With the help of a couple of `#ifdef`s, this pull request makes the `tcpinfo` field in `wangle::TransportInfo` available on macOS as well.

To further simplify getting the TCP state through `tcpinfo.tcpi_state` (which is my main motivation for doing this), I've added `#ifdef`s for the macOS values `TCPS_*` to the Linux/FreeBSD `TCP_*` macros.
Closes https://github.com/facebook/wangle/pull/125

Reviewed By: mzlee

Differential Revision: D7425728

Pulled By: yfeldblum

fbshipit-source-id: d5a93a516af5c3f8849b09d7816feb17bcffadce