Fix a broken test.
Summary:
D8986246 did move acception of an incoming connection to an earlier iteration
of event base loop. Previously we did not get a change to create read pipeline,
now we do. The fix is to adjust the expactations in the test.
Notice that this test only passes in opt mode and not in dev mode, due to ASAN
violations. I am not fixing this since the condition is pre-existing.

Reviewed By: hannesroth

Differential Revision: D9212420

fbshipit-source-id: 56c20a4173d71c6ad60bca10490321ebe47ea7c5