Do not detach transport on error in EvbHandshakeHelper.
Summary: This is not safe to do in general since we could have an async action outstanding. Moving the failed connection back to the main thread is not very useful anyway (other than potentially giving access to more information, but the full exception is already available). We can instead destroy the transport on the handshake evb.

Reviewed By: mingtaoy

Differential Revision: D9813326

fbshipit-source-id: 2672ad9e47a7d30468728444419244a14a0819d2