Executor running mode in LRUPersistentCache
Summary:
Currently LRUPersistentCache starts another thread to do all the sync
tasks. For one process has multiple such caches, we start a few threads. Each
such thread only do very small amount of work. This diff adds another way of
running these sync tasks on a shared executor. The diff also keeps the current
thread based approach untouched so that option will continue to work. Only when
an executor is passed to the cache we will use the executor approach.

When sync task is dispatched to executor, a weak_ptr to the cache itself is captured. So if the cache itself is going away, the task will be skipped. This is easier than the current stopSyncer_ flag passed and synced around cache destructor and the sync task.

In the current thread running mode, when the cache is going away, if the memory cache has been updated, we have one last chance to sync it to the persistence layer. In current diff, I'm not doing that, but it shouldn't be hard to change.

Reviewed By: mingtaoy

Differential Revision: D9436259

fbshipit-source-id: ac4baa4833cc47e2d3040cbafe7dbcd632699a8f