Move persistence loading onto the executor or sync thread in LRUPersistentCache
Summary:
Loading the persistence layer into memory right inline in the caller
thread of the LRUPersistentCache constructor could block the caller thread for
some amount of time. This diff moves the persistence loading on the executor
or the sync thread to unblock the caller thread.

With this change, the order between persistence loading and any operation on
the in-memory cache is no longer guranteed. But any operation on the in-memory
cache that happens before the persistence loading will be overriden by the
loading operation. Thus a condition variable is added to synchronize the
loading and all the operations on the in-memory cache. Besides the loadData
call, any other access to the in-memory cache is replaced with a blocking
function that waits for the loading to finish. So if by the time the user
put/get/clear/remove on the cache the persistence loading is still not
finished, we will still be blocked. But this opens the door for the user to
race file persistence cache reading and network IO.

Reviewed By: JunqiWang

Differential Revision: D13144621

fbshipit-source-id: f0f336c04b2bbe57ac54c0e0d4beb9ed270275a3