Remove syncOnDestroy setting in LRUPersistentCache
Summary:
The LRUPersistentCache will sync one last time upon destroy even
without this setting. I'm keeping the test case to ensure it's indeed the case.
Also the test was flaky since the number of times persis_ gets called in current
code depends on if put() happen before the first sync on sync thread or after
that due to a version mismatch bug between persistence layer and in-memory
cache layer. So i also changed the test case. The version mismatch bug is being
addressed in another diff.

Reviewed By: JunqiWang

Differential Revision: D13280808

fbshipit-source-id: 1c29ba25b481b0fa1cc20e185d3345b8826c87f6