-cleanup-context-updates
Summary:
Currently the way we update the fizz server context for tls ticket seed or ssl context updates is pretty messy.
The current approach is based on the idea that we shouldn't be updating the server context after it's set on the acceptor. This allows us to avoid races and makes things simpler. To do this we store the cipher + cert manager on the acceptor itself, so that we aren't recreating these on any update.

Reviewed By: knekritz

Differential Revision: D10528367

fbshipit-source-id: 40a4741662b5e97ce9d08c06c110ea14e177929e