Sync LRUPersistentCache upon destroy when it's running in executor mode
Summary:
In the thread running mode, when LRUPersistentCache is destructed, we
do one last sync to bring any change of the in-memory cache to the persistence
layer. In executor mode, however, no such sync will happen when the cache is
destructed. This could potentially lead to critical settings not getting stored
on the client. This diff adds the syncOnDestroy behavior for executor running
mode.

Reviewed By: elindsey

Differential Revision: D13162300

fbshipit-source-id: 4422fd8d5b8d4b7f7f16ea91a76316db059f2931