Fix race in LRUPersistentCache setPersistence
Summary:
When I changed the setPersistenceHelper to use a SaturatingSemaphore
to make it only happen once, I completely missed the case where someone can
replace the persistence in the middle of a sync via the rogue setPersistence()
API. So when one thread setPersistence(), and reset()s the SaturatingSemaphore, if
the sync thread is in setPersistenceHelper in the meantime, it will find the
semaphore not ready(), continue to load the persistence. And the thread that
calls setPersistence() will also find the sempahore not ready(), double load
the persistence.

This diff moves the semaphore checking behind the persistence write lock as
well. So at any moment only one thread can find the semaphore not ready(), load
it, and post() the semaphore.

Reviewed By: fugalh

Differential Revision: D13469774

fbshipit-source-id: 4d70d791e5892dbab20da708efe528cfcefdec93