Properly handle freeing of SSL_SESSION when removing from local cache
Summary: The current implementation of the local cache will leak sessions when sessions are explicitly removed from the cache. This wouldn't happen in 1.0.2 due to gating around the callback that changed in 1.1.0. With 1.1.0, session removal callbacks are now being called, and the code fails to call SSL_SESSION_free on the sessions before deleting them (LRUCacheMap only frees sessions it prunes, not replacements or explicit removals)

Reviewed By: siyengar

Differential Revision: D13668226

fbshipit-source-id: 52ac6e990b964cec00d19772a9e1bcfcb1f49269