wangle: really fix FilePoller on win32
Summary:
We've made a couple of different attempts at resolving this,
but are still having problems because the version of MSVC 2017 installed
on some of our CI systems doesn't have a fully baked `std::filesystem`
implementation.

This diff resolves the situation by making the problematic nanoseconds
portion of the timestamp posix specific.

Reviewed By: yfeldblum

Differential Revision: D14697284

fbshipit-source-id: 8733ffc18235089e2127b7871e0b405261d3cedf