wangle: remove std::filesystem related code
Summary:
our nemesis, `std::filesystem`.  On this gcc 8 system the headers support
`std::filesystem` but the implementation requires linking with `-lstc++fs` and
we don't have any cmake support for figuring this out, resulting in linker
failures when building anything that relies on the FilePoller class.

As of my recent change to fix the build for that on windows, we don't really
need `std::filesystem` at all, so I'd be in favor of removing that code on
the basis that less code is preferable to more code.

Reviewed By: yfeldblum

Differential Revision: D14725616

fbshipit-source-id: 1988476ace2fc3494bbead52f9995ec1f5fbc95f