Capture session byte offsets for first and last body bytes
Summary:
Adds fields for the session byte offset of the first and last bytes of a response body.

Protocols that support preemption and multiplexing (e.g., HTTP/2) may write multiple response body in parallel to the transport. Capturing the first and last body byte offsets enables examination of this multiplexing.

The difference between these two offsets is also useful for measuring throughput as it provides the total number of bytes transferred via transport between the time the first byte of the response was flushed (`TransportInfo::timeToFirstByte`) and when the ack was received for the last byte in the response (`TransportInfo::timeToLastBodyByteAck`).

Reviewed By: udippant

Differential Revision: D14731330

fbshipit-source-id: 12fa21ca68f22e9a57ad0c20987da0d7f3c9e1e3