Allow socket options to be based into BaseClientBootstrap/ClientBootstrap
Summary:
This allows socket options to be passed into the client bootstrap and used during the connection.

This makes it so that users of the basic ClientBootstrap implementation can add socket options when necessary. This makes it possible to easily add things like IP TOS options without having to write a bunch of boilerplate code to override the ClientBootstrap implementation if that's all that's needed.

Reviewed By: yfeldblum

Differential Revision: D18893686

fbshipit-source-id: 096e4aed4426d505e7b1523d8e7be7528d281484