Pull default SSL context out of SslContexts and into SSLContextManager
Summary: This makes it so the SslContexts struct no longer holds a reference to the default context, instead storing the key data that would normally point to it inside the map. This way, the SslContexts struct can be referred to from within the default context (specifically, its server name callback).

Reviewed By: mingtaoy

Differential Revision: D19156168

fbshipit-source-id: 1629f3ed3f0611cadd8f3013492e9e9ee642ff56