collectX to collectXSemiFuture or collectXUnsafe
Summary:
Migration from Future-returning executor-erasing collectX forms to
SemiFuture-returning forms, that are less risky in particular with coroutines.

Earlier changes added collectXSemiFuture and collectXUnsafe as a migration
path. This diff is part of a migration off of the basic collectX forms where
there might be some conflict with the change of return type in Phase 4.

Reviewed By: yfeldblum

Differential Revision: D20631253

fbshipit-source-id: 74c2ca76f66f59facaa7cc8d83f6f3caf67daadf