OpenSSL callback return code bug-fix
Summary:
Currently, when the `tlsext_ticket_key_cb` can't find ticket secrets in the
`TLSTicketKeyManager` on the ticket encrypt codepath we return -1, which is a
fatal error.

This works right now because we only set the callback inside the ticket key
manager, and there's a branch that sets the callback to null if there aren't
any encryption keys available. An upcoming diff allows us to construct a
ticket key manager then explicitly attach it to the context, meaning we want to
unconditionally set the callback and make it the user of this class's
responsibility to ensure the keys have been configured correctly.

The net result is that we can no longer consider it a fatal error if the
callback is invoked prior to keys being available and instead we simply continue
the connection without issuing a new ticket.

Reviewed By: mingtaoy

Differential Revision: D26343618

fbshipit-source-id: 5338eaa5f22641c9ea84bad781530d6580e9e0d6