update the branch name for the fb303-source manifest
Summary:
The primary branch in the fb303 github repository was renamed from `master` to
`main`.  Update the `fb303-source` manifest to reflect this.

The main `fb303` manifest was already updated in D30700180.  The
`fb303-source` manifest is a separate manifest that only fetches the fb303
source code without building it.  This is only used by the `eden_scm`
manifest.

Reviewed By: fanzeyi

Differential Revision: D30794650

fbshipit-source-id: 0e7790f8ceb737ce2f26ab213e1f0f94c5b3f045