Fix MultiFilePoller deadlock
Summary:
Currently there's a lock order inversion bug for the FilePoller lock and the
MultiFilePoller lock.

Registering files and cancelling callbacks both attempt to acquire the
FilePoller lock while holding the MultiFilePoller lock.

When a file is changed, we attempt to acquire the MultiFilePoller lock while
holding the FilePoller lock.

To fix this, we now maintain the invariant that we never attempt to acquire the
FilePoller lock while holding the MultiFilePoller lock.

This means that while registering files, we might be in an intermediate state
where the filepath(s)->callback(s) mapping might be fully up to date, but we
haven't yet started watching the necessary files.

Similarly, while deleting callbacks we might be in an intermediate state where
the filepath(s)->callback(s) mapping might be fully up to date but we are still
unnecessarily watching a file we no longer need to (and if the file(s) is/are
modified in that time the callbacks become no-ops.

Both of these seem like perfectly reasonable states to be in. A more invasive
refactor might involve removing the ability to dynamically add and remove
callbacks once the filepoller has started, but this is sufficient for now.

Reviewed By: yfeldblum

Differential Revision: D31182613

fbshipit-source-id: 73c94e3d066a47d8b274ae34ee750313737efef9