Add folly::RequestContext support to Pipeline
Summary:
`folly::RequestContext` can be used to store request specific metrics such as CPU and memory usage. This RequestContext can be automatically passed to other threads, so it's recommended to use a folly::RequestContextScopeGuard whenever storing any info in a RequestContext.

This should be a no-op unless RequestContext is explicitly set

Reviewed By: yfeldblum

Differential Revision: D31149532

fbshipit-source-id: 008bba16cff15f97d059bd4169bfec105c1a9fa1