Fixing sign loss in ConnectionManager
Summary:
I accidentally found few lines in the code where `std::max` is used with explicit template specification: `std::max<size_t>(0, ...)`. Obviously, the purpose of such an expression was to return zero instead of a negative value. But with the explicit specialization by unsigned type, this expression loses its meaning, because a negative integer will be converted to an unsigned one, and then a comparison with zero will be made.

If we do not take into account explicit type specification, then it seems like the author wanted to limit the `pct` variable to the segment [0, 1].

Reviewed By: yfeldblum, ot, luciang

Differential Revision: D31741241

fbshipit-source-id: a3a5561a9e004056b25232376c892dd97ff25d87