Add initial support for fizz kTLS
Summary:
This diff introduces a `preferKTLS` option in wangle's Fizz settings, such that
an acceptor configured with this flag enabled will attempt to upgrade accepted
Fizz sockets into the kTLS version before passing it to the application.

`preferKTLS` is a wangle managed setting. Consequently, the upgrade is performed
as part of a non-overridable `transformTransport` function which subclasses
cannot alter.

Reviewed By: anhuang

Differential Revision: D30704292

fbshipit-source-id: b7c354418a9d33822f8be2309d15e34879e3ec77