Lift FizzAcceptorHandshakeHelper::LoggingCallback to wangle namespace
Summary:
This will make more sense in context with the later diffs in this stack.

This diff renames wangle::FizzAcceptorHandshakeHelper::LoggingCallback to
wangle::FizzLoggingCallback. Additionally, it adjusts the logging callback
interface to expect a reference to a wangle::TransportInfo, since it more
accurately represents the fact that it will never be null.

Reviewed By: anhuang

Differential Revision: D31705282

fbshipit-source-id: a2cfcfa364001c4daf595570a20abb74c356a5d1